+++ apps/frontend/src/modules/chat/chat-api.ts
@@ -4,6 +4,7 @@ import { type APIConfiguration } from "~/libs/types/types.js";
 
 import { AIAssistantApiPath } from "./libs/enums/enums.js";
 import {
+	type AIAssistantChangeTaskRequestDto,
 	type AIAssistantChatInitializeResponseDto,
 	type AIAssistantCreateMultipleTasksDto,
 	type AIAssistantResponseDto,
@@ -13,6 +14,22 @@ import {
 class ChatApi extends BaseHTTPApi {
 	public constructor({ baseUrl, http, storage }: APIConfiguration) {
 		super({ baseUrl, http, path: APIPath.ASSISTANT, storage });
+	}
+
+	public async changeTasksSuggestion(
+		payload: AIAssistantChangeTaskRequestDto,
+	): Promise<AIAssistantResponseDto> {
+		const response = await this.load(
+			this.getFullEndpoint(AIAssistantApiPath.CHAT_CHANGE_TASKS, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "POST",
+				payload: JSON.stringify(payload),
+			},
+		);
+
+		return await response.json<AIAssistantResponseDto>();
 	}
 
 	public async createTasksFromSuggestions(