+++ apps/frontend/src/pages/chat/libs/components/buttons-controller/libs/components/dislike-suggetsions-options/dislike-suggestions-options.tsx
@@ -1,0 +1,68 @@
+import { Button, Icon } from "~/libs/components/components.js";
+import {
+	useAppDispatch,
+	useAppSelector,
+	useCallback,
+} from "~/libs/hooks/hooks.js";
+import { actions as chatActions } from "~/modules/chat/chat.js";
+import { ButtonsModeOption } from "~/pages/chat/libs/enums/enums.js";
+
+import { RegenerateSuggestionButton } from "./libs/components/components.js";
+import { DISLIKE_ALL_SUGGESTIONS_BUTTON_LABEL } from "./libs/constants/constants.js";
+import { DislikeSuggestionMessage } from "./libs/enums/enums.js";
+import styles from "./styles.module.css";
+
+const DislikeSuggestionsOptions: React.FC = () => {
+	const { taskSuggestions } = useAppSelector((state) => ({
+		taskSuggestions: state.chat.taskSuggestions,
+	}));
+	const dispatch = useAppDispatch();
+
+	const handleRegenerateAllSuggestions = useCallback(() => {
+		dispatch(
+			chatActions.addAssistantTextMessage(DislikeSuggestionMessage.MAIN),
+		);
+		dispatch(
+			chatActions.addUserTextMessage(DISLIKE_ALL_SUGGESTIONS_BUTTON_LABEL),
+		);
+		dispatch(
+			chatActions.addAssistantTextMessage(DislikeSuggestionMessage.WAIT),
+		);
+		dispatch(chatActions.setButtonsMode(ButtonsModeOption.NONE));
+		void dispatch(
+			chatActions.changeTasksSuggestion({
+				APIPayload: { messages: [], tasks: taskSuggestions },
+			}),
+		);
+	}, [dispatch, taskSuggestions]);
+
+	return (
+		<div className={styles["message-container"]}>
+			<Icon name="aiAssistantAvatar" />
+			<div className={styles["content-container"]}>
+				<div className={styles["content"]}>
+					<p>{DislikeSuggestionMessage.MAIN}</p>
+					<div className={styles["button-container"]}>
+						<Button
+							label={DISLIKE_ALL_SUGGESTIONS_BUTTON_LABEL}
+							onClick={handleRegenerateAllSuggestions}
+							variant="secondary"
+						/>
+						{taskSuggestions.map((suggestion) => {
+							return (
+								<RegenerateSuggestionButton
+									key={suggestion.categoryId}
+									label={`${suggestion.categoryName} sector task`}
+									oldSuggestions={taskSuggestions}
+									suggestion={suggestion}
+								/>
+							);
+						})}
+					</div>
+				</div>
+			</div>
+		</div>
+	);
+};
+
+export { DislikeSuggestionsOptions };