+++ apps/mobile/src/packages/chat/chat-api.ts
@@ -3,11 +3,12 @@ import { type APIConfiguration, BaseHttpApi } from "~/libs/packages/api/api";
 
 import { AIAssistantApiPath } from "./libs/enums/enums";
 import {
+	type AIAssistantChangeTaskRequestDto,
+	type AIAssistantChatInitializeResponseDto,
 	type AIAssistantCreateMultipleTasksDto,
-	type AIAssistantRequestDto,
+	type AIAssistantExplainTaskRequestDto,
 	type AIAssistantResponseDto,
 	type AIAssistantSuggestTaskRequestDto,
-	type ThreadMessageCreateDto,
 } from "./libs/types/types";
 
 class ChatApi extends BaseHttpApi {
@@ -32,10 +33,26 @@ class ChatApi extends BaseHttpApi {
 	}
 
 	public async getChangedTasksSuggestion(
-		payload: AIAssistantRequestDto,
+		payload: AIAssistantChangeTaskRequestDto,
 	): Promise<AIAssistantResponseDto> {
 		const response = await this.load(
 			this.getFullEndpoint(AIAssistantApiPath.CHAT_CHANGE_TASKS, {}),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "POST",
+				payload: JSON.stringify(payload),
+			},
+		);
+
+		return await response.json<AIAssistantResponseDto>();
+	}
+
+	public async getExplainedTasksSuggestion(
+		payload: AIAssistantExplainTaskRequestDto,
+	): Promise<AIAssistantResponseDto> {
+		const response = await this.load(
+			this.getFullEndpoint(AIAssistantApiPath.CHAT_EXPLAIN_TASKS, {}),
 			{
 				contentType: ContentType.JSON,
 				hasAuth: true,
@@ -63,7 +80,7 @@ class ChatApi extends BaseHttpApi {
 		return await response.json<AIAssistantResponseDto>();
 	}
 
-	public async initializeConversation(): Promise<ThreadMessageCreateDto> {
+	public async initializeConversation(): Promise<AIAssistantChatInitializeResponseDto> {
 		const response = await this.load(
 			this.getFullEndpoint(AIAssistantApiPath.CHAT_INITIALIZE, {}),
 			{
@@ -74,7 +91,7 @@ class ChatApi extends BaseHttpApi {
 			},
 		);
 
-		return await response.json<ThreadMessageCreateDto>();
+		return await response.json<AIAssistantChatInitializeResponseDto>();
 	}
 }