+++ apps/mobile/src/packages/chat/chat.ts
@@ -17,10 +17,16 @@ export {
 	ChatMessageAuthor,
 	ChatMessageType,
 } from "./libs/enums/enums";
-export { checkIsTaskType, processMessages } from "./libs/helpers/helpers";
 export {
+	checkIsTaskType,
+	processExplainedTasksMessages,
+	processMessages,
+} from "./libs/helpers/helpers";
+export {
+	type AIAssistantChangeTaskRequestDto,
+	type AIAssistantChatInitializeResponseDto,
 	type AIAssistantCreateMultipleTasksDto,
-	type AIAssistantRequestDto,
+	type AIAssistantExplainTaskRequestDto,
 	type AIAssistantResponseDto,
 	type AIAssistantSuggestTaskRequestDto,
 	type ChatMessage,