+++ apps/mobile/src/packages/chat/libs/types/types.ts
@@ -1,7 +1,9 @@
 export { type ChatMessage } from "./chat-message.type";
 export {
+	type AIAssistantChangeTaskRequestDto,
+	type AIAssistantChatInitializeResponseDto,
 	type AIAssistantCreateMultipleTasksDto,
-	type AIAssistantRequestDto,
+	type AIAssistantExplainTaskRequestDto,
 	type AIAssistantResponseDto,
 	type AIAssistantSuggestTaskRequestDto,
 	type ChatMessageDto,