+++ apps/mobile/src/screens/chat/libs/components/chat-buttons/libs/components/accept-regenerate-task-buttons/accept-regenerate-task-buttons.tsx
@@ -13,65 +13,66 @@ import { type NativeStackNavigationProp } from "~/libs/types/types";
 import { ButtonsMode, ChatMessageAuthor } from "~/packages/chat/chat";
 import { actions as chatActions } from "~/slices/chat/chat";
 
-import { FEEDBACK_TEXT } from "./constants/constants";
-import { type RootStackParameterList } from "./types/types";
+import { AI_RESPONSE, EXPLAIN_TEXT } from "../../constants/constants";
+import { type RootStackParameterList } from "../../types/types";
 
-const AcceptTasks: React.FC = () => {
+const AcceptRegenerateTaskButtons: React.FC = () => {
 	const dispatch = useAppDispatch();
 	const navigation =
 		useNavigation<NativeStackNavigationProp<RootStackParameterList>>();
 
-	const threadId = useAppSelector((state) => state.chat.threadId);
 	const taskSuggestions = useAppSelector((state) => state.chat.taskSuggestions);
 
 	const handleAccept = useCallback(() => {
 		void dispatch(
 			chatActions.createTasks({
-				payload: taskSuggestions,
-				threadId: threadId as string,
+				messages: [],
+				tasks: taskSuggestions,
 			}),
 		);
 
 		dispatch(chatActions.setButtonsMode(ButtonsMode.NONE));
-		dispatch(
-			chatActions.addTextMessage({
-				author: ChatMessageAuthor.ASSISTANT,
-				text: FEEDBACK_TEXT,
-			}),
-		);
+
 		dispatch(
 			chatActions.addTextMessage({
 				author: ChatMessageAuthor.USER,
-				text: "Accept",
+				text: "Everything is clear",
+			}),
+		);
+
+		dispatch(
+			chatActions.addTextMessage({
+				author: ChatMessageAuthor.ASSISTANT,
+				text: AI_RESPONSE,
 			}),
 		);
 
 		navigation.navigate(BottomTabScreenName.TASKS);
-	}, [dispatch, taskSuggestions, threadId, navigation]);
+	}, [dispatch, taskSuggestions, navigation]);
 
 	const handleGenerateNew = useCallback(() => {
 		void dispatch(
 			chatActions.getChangedTasksSuggestion({
-				payload: taskSuggestions,
-				threadId: threadId as string,
+				messages: [],
+				tasks: taskSuggestions,
 			}),
 		);
 
 		dispatch(
 			chatActions.addTextMessage({
 				author: ChatMessageAuthor.USER,
-				text: "Generate new",
+				text: "I don't like the tasks.",
 			}),
 		);
-	}, [dispatch, taskSuggestions, threadId]);
+	}, [dispatch, taskSuggestions]);
 
 	return (
-		<ChatMessage text={FEEDBACK_TEXT}>
+		<ChatMessage text={EXPLAIN_TEXT}>
 			<View style={[globalStyles.gap12, globalStyles.mt8]}>
-				<Button label="Accept" onPress={handleAccept} />
+				<Button label="Everything is clear" onPress={handleAccept} />
 				<Button
 					appearance="outlined"
-					label="Regenerate new"
+					label="I don't like the tasks"
 					onPress={handleGenerateNew}
 				/>
 			</View>
@@ -79,4 +80,4 @@ const AcceptTasks: React.FC = () => {
 	);
 };
 
-export { AcceptTasks };
+export { AcceptRegenerateTaskButtons };