+++ apps/mobile/src/slices/chat/actions.ts
@@ -2,19 +2,21 @@ import { createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type AsyncThunkConfig } from "~/libs/types/types";
 import {
+	type AIAssistantChangeTaskRequestDto,
+	type AIAssistantChatInitializeResponseDto,
 	type AIAssistantCreateMultipleTasksDto,
-	type AIAssistantRequestDto,
+	type AIAssistantExplainTaskRequestDto,
 	type AIAssistantSuggestTaskRequestDto,
 	type ChatMessage,
+	processExplainedTasksMessages,
 	processMessages,
 	type TaskCreateDto,
-	type ThreadMessageCreateDto,
 } from "~/packages/chat/chat";
 
 import { name as sliceName } from "./chat.slice";
 
 const initConversation = createAsyncThunk<
-	ThreadMessageCreateDto,
+	AIAssistantChatInitializeResponseDto,
 	undefined,
 	AsyncThunkConfig
 >(`${sliceName}/initialize-conversation`, async (_, { extra }) => {
@@ -42,13 +44,27 @@ const getChangedTasksSuggestion = createAsyncThunk<
 		messages: ChatMessage[];
 		taskSuggestions: TaskCreateDto[];
 	},
-	AIAssistantRequestDto,
+	AIAssistantChangeTaskRequestDto,
 	AsyncThunkConfig
 >(`${sliceName}/get-changed-tasks-suggestion`, async (payload, { extra }) => {
 	const { chatApi } = extra;
 	const response = await chatApi.getChangedTasksSuggestion(payload);
 
 	return processMessages(response.messages);
+});
+
+const getExplainedTasksSuggestion = createAsyncThunk<
+	{
+		messages: ChatMessage[];
+		taskExplanations: TaskCreateDto[];
+	},
+	AIAssistantExplainTaskRequestDto,
+	AsyncThunkConfig
+>(`${sliceName}/get-explained-tasks-suggestion`, async (payload, { extra }) => {
+	const { chatApi } = extra;
+	const response = await chatApi.getExplainedTasksSuggestion(payload);
+
+	return processExplainedTasksMessages(response.messages);
 });
 
 const createTasks = createAsyncThunk<
@@ -64,6 +80,7 @@ const createTasks = createAsyncThunk<
 export {
 	createTasks,
 	getChangedTasksSuggestion,
+	getExplainedTasksSuggestion,
 	getTasksForCategories,
 	initConversation,
 };