+++ packages/shared/src/modules/chats/chats.ts
@@ -1,8 +1,11 @@
 export { ChatMessageAuthor, ChatMessageType } from "./libs/enums/enums.js";
 export { checkIsTaskType } from "./libs/helpers/helpers.js";
 export {
+	type ChangeTasksSuggestionPayload,
+	type ChatMessage,
 	type ChatMessageCreateDto,
 	type ChatMessageDto,
+	type ProcessedMessagesAndSuggestions,
 	type TaskMessage,
 	type TextMessage,
 } from "./libs/types/types.js";